from_52/src/requirements.txt
colorama==0.4.1
colorlog==4.0.2
Django==2.2.2
djangorestframework==3.9.4
numpy==1.18.0
opencv-python==4.1.1.26
Pillow==6.0.0
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
